Move aabb.h to common/
