--- src/TheNextWeek/aabb.h
@@ -1,56 +1,0 @@
-#ifndef AABB_H
-#define AABB_H
-//==============================================================================================
-// Originally written in 2016 by Peter Shirley <ptrshrl@gmail.com>
-//
-// To the extent possible under law, the author(s) have dedicated all copyright and related and
-// neighboring rights to this software to the public domain worldwide. This software is
-// distributed without any warranty.
-//
-// You should have received a copy (see file COPYING.txt) of the CC0 Public Domain Dedication
-// along with this software. If not, see <http://creativecommons.org/publicdomain/zero/1.0/>.
-//==============================================================================================
-
-#include "common/rtweekend.h"
-
-
-inline double ffmin(double a, double b) { return a < b ? a : b; }
-inline double ffmax(double a, double b) { return a > b ? a : b; }
-
-class aabb {
-    public:
-        aabb() {}
-        aabb(const vec3& a, const vec3& b) { _min = a; _max = b; }
-
-        vec3 min() const {return _min; }
-        vec3 max() const {return _max; }
-
-        bool hit(const ray& r, double tmin, double tmax) const {
-            for (int a = 0; a < 3; a++) {
-                auto t0 = ffmin((_min[a] - r.origin()[a]) / r.direction()[a],
-                                (_max[a] - r.origin()[a]) / r.direction()[a]);
-                auto t1 = ffmax((_min[a] - r.origin()[a]) / r.direction()[a],
-                                (_max[a] - r.origin()[a]) / r.direction()[a]);
-                tmin = ffmax(t0, tmin);
-                tmax = ffmin(t1, tmax);
-                if (tmax <= tmin)
-                    return false;
-            }
-            return true;
-        }
-
-        vec3 _min;
-        vec3 _max;
-};
-
-aabb surrounding_box(aabb box0, aabb box1) {
-    vec3 small( ffmin(box0.min().x(), box1.min().x()),
-                ffmin(box0.min().y(), box1.min().y()),
-                ffmin(box0.min().z(), box1.min().z()));
-    vec3 big  ( ffmax(box0.max().x(), box1.max().x()),
-                ffmax(box0.max().y(), box1.max().y()),
-                ffmax(box0.max().z(), box1.max().z()));
-    return aabb(small,big);
-}
-
-#endif
